Organize classes in the starpu directory. Update documentation
