+++ include/tlapack/starpu/types.hpp
@@ -1,0 +1,24 @@
+/// @file starpu/types.hpp
+/// @brief Types for StarPU.
+/// @author Weslley S Pereira, University of Colorado Denver, USA
+//
+// Copyright (c) 2021-2023, University of Colorado Denver. All rights reserved.
+//
+// This file is part of <T>LAPACK.
+// <T>LAPACK is free software: you can redistribute it and/or modify it under
+// the terms of the BSD 3-Clause license. See the accompanying LICENSE file.
+
+#ifndef TLAPACK_STARPU_TYPES_HH
+#define TLAPACK_STARPU_TYPES_HH
+
+#include <cstdint>
+
+namespace tlapack {
+namespace starpu {
+
+    using idx_t = uint32_t;
+
+}  // namespace starpu
+}  // namespace tlapack
+
+#endif  // TLAPACK_STARPU_TYPES_HH
